# BE_FARMX/openapi.yaml
openapi: 3.0.0
info:
  title: FarmX API
  description: API for Farm/Pond/Scan Management System
  version: 1.0.0
  contact:
    name: FarmX Team
servers:
  - url: http://localhost:4000/api
    description: Development server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ScanResult:
      type: object
      properties:
        _id:
          type: string
        pond:
          type: string
        deviceId:
          type: string
        healthScore:
          type: number
          minimum: 0
          maximum: 100
        diseasePrediction:
          type: object
          properties:
            disease:
              type: string
            confidence:
              type: number
            recommendations:
              type: array
              items:
                type: string
        metrics:
          type: object
          additionalProperties: true
        rawData:
          type: object
        imageUrl:
          type: string
          format: uri
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
paths:
...
  /scans:
    get:
      tags:
        - Scans
      summary: Get all scans
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: pondId
          schema:
            type: string
        - in: query
          name: page
          schema:
            type: integer
            default: 1
        - in: query
          name: limit
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: List of scans
    post:
      tags:
        - Scans
      summary: Create scan result
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - metrics
              properties:
                deviceId:
                  type: string
                  example: DEVICE-001
                metrics:
                  type: object
                  example:
                    weight: 250
                    length: 25.5
                    MGR: 1.5
                rawData:
                  type: object
                saveToPondId:
                  type: string
      responses:
        "201":
          description: Scan created
...
  /ponds/{pondId}/scans:
    get:
      tags:
        - Scans
      summary: Get all scans for a pond
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: pondId
          required: true
          schema:
            type: string
        - in: query
          name: page
          schema:
            type: integer
            default: 1
        - in: query
          name: limit
          schema:
            type: integer
            default: 10
      responses:
        "200":
          description: List of scans
  /scans/history:
    get:
      tags:
        - Scans
      summary: Get scan history
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: pondId
          schema:
            type: string
        - in: query
          name: page
          schema:
            type: integer
            default: 1
        - in: query
          name: limit
          schema:
            type: integer
            default: 50
      responses:
        "200":
          description: Scan history
  /scans/monthly-stats:
    get:
      tags:
        - Scans
      summary: Get monthly scan statistics
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: month
          schema:
            type: integer
            minimum: 1
            maximum: 12
        - in: query
          name: year
          schema:
            type: integer
      responses:
        "200":
          description: Monthly statistics
  /scans/{id}:
    get:
      tags:
        - Scans
      summary: Get a single scan result
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Scan details
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  scan:
                    $ref: "#/components/schemas/ScanResult"
    put:
      tags:
        - Scans
      summary: Update a scan result
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                metrics:
                  type: object
                healthScore:
                  type: number
                  minimum: 0
                  maximum: 100
                diseasePrediction:
                  type: object
                imageUrl:
                  type: string
                  format: uri
                saveToPondId:
                  type: string
      responses:
        "200":
          description: Scan updated
    delete:
      tags:
        - Scans
      summary: Delete a scan result
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Scan deleted
...